requirements.txt
django==3.0.5
djangorestframework==3.12.4
python-dotenv==0.17.1
gunicorn==20.0.4         
psycopg2-binary==2.8.5
django-jsonfield==1.4.1 
django-filter==2.4.0